Move search code into new package
